Moved config utilities to settings module.
